push changes to gh-pages

+++ www/js/components/accounts/manage-accounts.directive.js
@@ -11,8 +11,9 @@ angular
 	.directive('manageAccounts',[
 		'accountService',
 		'alertService',
+		'cleanupService',
 		'$state',
-		function(accountService, alertService, $state) {
+		function(accountService, alertService, cleanupService, $state) {
 		return {
 			restrict: 'E',
 			templateUrl: 'templates/components/accounts/manage-accounts.template.html',
@@ -26,8 +27,11 @@ angular
 					if (response) {
 						if (accountService.isUnique(response.loginid)) {
 							accountService.add(response);
+							accountService.setDefault(response.token);
 							scope.accounts = accountService.getAll();
-							scope.$apply();
+							if(!scope.$$phase){
+								scope.$apply();
+							}
 						} else {
 							if (scope.settingDefault) {
 								scope.settingDefault = false;
@@ -44,7 +48,9 @@ angular
 				scope.$on('token:remove', function(e, response) {
 					accountService.remove(response);
 					scope.accounts = accountService.getAll();
-					scope.$apply();
+					if(!scope.$$phase){
+						scope.$apply();
+					}
 				});
 
 				scope.addAccount = function(_token) {
@@ -70,8 +76,23 @@ angular
 				};
 
 				scope.removeAllAccounts = function() {
-					accountService.removeAll();
-					$state.go('signin');
+					alertService.confirmRemoveAllAccount(
+						function(res){
+							if(typeof(res) !== "boolean"){
+								if(res == 1)
+									res = true;
+								else
+									res = false;
+							}
+
+							if(res){
+								accountService.removeAll();
+								cleanupService.run();
+								$state.go('signin');
+							}
+						}
+					);
+					
 				};
 			}
 		};
